Move helper functions to chunks.ts
