+++ src/loaders/OmeZarrLoader.ts
@@ -35,7 +35,8 @@ import { VolumeLoadError, VolumeLoadErrorType, wrapVolumeLoadError } from "./Vol
 import { relaxedFetch, withVoleInstrumentation } from "./zarr_utils/wrappers.js";
 import { assertMetadataHasMultiscales, toOMEZarrMetaV4, validateOMEZarrMetadata } from "./zarr_utils/validation.js";
 import { remapUri } from "../utils/url_utils.js";
-import { ARRAY_CONSTRUCTORS, type TypedArray } from "../types.js";
+import { type TypedArray } from "../types.js";
+import { getChunkByteSize, iterateChunkCoords } from "./zarr_utils/chunks.js";
 
 const CHUNK_REQUEST_CANCEL_REASON = "chunk request cancelled";
 
@@ -91,39 +92,6 @@ const DEFAULT_FETCH_OPTIONS = {
   maxPrefetchDistance: [12, 5, 5, 5] as [number, number, number, number],
   maxPrefetchChunks: 30,
 };
-
-function getChunkByteSize(level: NumericZarrArray): number {
-  const voxelsPerChunk = level.chunks.reduce((accumulator, chunkSize) => accumulator * chunkSize, 1);
-  return voxelsPerChunk * ARRAY_CONSTRUCTORS[level.dtype].BYTES_PER_ELEMENT;
-}
-
-/**
- * @param level One resolution level of a ZARR
- * @yields One coords array per chunk in the level
- */
-function* iterateChunkCoords(level: NumericZarrArray): Generator<number[]> {
-  // level.chunks[dimension] is the width of a chunk in the given dimenion
-  const chunksPerDimension = level.shape.map((size, dimension) => Math.ceil(size / level.chunks[dimension]));
-
-  /**
-   * @param coords A partial coordinates array (e.g., [1,2,3])
-   * @yields Coordinates arrays that start with the values in coords (e.g., [1,2,3,4,5])
-   */
-  function* iterateCoordsWithPrefix(coords: number[]): Generator<number[]> {
-    const nextDimension = coords.length;
-    if (nextDimension === chunksPerDimension.length) {
-      // coords is fully filled out with values in all dimensions
-      yield coords;
-      return;
-    }
-
-    for (let thisDimIdx = 0; thisDimIdx < chunksPerDimension[nextDimension]; thisDimIdx++) {
-      yield* iterateCoordsWithPrefix([...coords, thisDimIdx]);
-    }
-  }
-
-  yield* iterateCoordsWithPrefix([]);
-}
 
 class OMEZarrLoader extends ThreadableVolumeLoader {
   /** The ID of the subscriber responsible for "actual loads" (non-prefetch requests) */
